ActionStepV2 refactor
