+++ frontend/src/scenes/actions/Actions.scss
@@ -58,7 +58,7 @@
     }
 }
 
-// ActionStepV2.js
+// ActionStep
 .action-step {
     margin-bottom: 0;
     height: 100%;
